+++ frontend/src/modules/event-logs/EventLogs.tsx
